requirements/base.txt
Django==1.11.7
pytz==2017.3
djangorestframework==3.7.3
psycopg2==2.7.3.2
coreapi==2.3.3
pygments==2.2.0
markdown==2.6.10
